requirements.txt
langchain
langchain-community
langchain-openai
langgraph
pydantic
pypdf
python-dotenv